feat: add component editor modal and settings tab
