+++ src/pages/FlowEditor/StepForm/Custom.tsx
@@ -4,6 +4,12 @@ import { Label } from "../../../components/ui/label";
 import { Textarea } from "../../../components/ui/textarea";
 import { Button } from "../../../components/ui/button";
 import {
+  Dialog,
+  DialogContent,
+  DialogHeader,
+  DialogTitle,
+} from "../../../components/ui/dialog";
+import {
   Select,
   SelectContent,
   SelectItem,
@@ -45,6 +51,7 @@ export default function CustomStepForm({ step, setField }: Props) {
   const { components, load, add, update } = useCustomComponents();
   const [current, setCurrent] = useState<CustomComponent | null>(null);
   const [isNew, setIsNew] = useState(false);
+  const [open, setOpen] = useState(false);
 
   useEffect(() => {
     load();
@@ -105,77 +112,94 @@ export default function CustomStepForm({ step, setField }: Props) {
       await update(sanitized);
       setCurrent(sanitized);
     }
+    setOpen(false);
   };
 
   return (
     <div className="space-y-4">
-      <div className="space-y-2">
-        <Label className="text-sm font-medium">Componente</Label>
-        <Select
-          value={step.componentId || (isNew ? NEW_VALUE : "")}
-          onValueChange={handleSelect}
-        >
-          <SelectTrigger>
-            <SelectValue placeholder="Escolha ou crie" />
-          </SelectTrigger>
-          <SelectContent>
-            {components.map((c) => (
-              <SelectItem key={c.id} value={c.id}>
-                {c.name}
-              </SelectItem>
-            ))}
-            <SelectItem value={NEW_VALUE}>Novo componente...</SelectItem>
-          </SelectContent>
-        </Select>
+      <div className="flex items-end gap-2">
+        <div className="flex-1 space-y-2">
+          <Label className="text-sm font-medium">Componente</Label>
+          <Select
+            value={step.componentId || (isNew ? NEW_VALUE : "")}
+            onValueChange={handleSelect}
+          >
+            <SelectTrigger>
+              <SelectValue placeholder="Escolha ou crie" />
+            </SelectTrigger>
+            <SelectContent>
+              {components.map((c) => (
+                <SelectItem key={c.id} value={c.id}>
+                  {c.name}
+                </SelectItem>
+              ))}
+              <SelectItem value={NEW_VALUE}>Novo componente...</SelectItem>
+            </SelectContent>
+          </Select>
+        </div>
+        <Button variant="outline" size="sm" onClick={() => setOpen(true)}>
+          {isNew || !step.componentId ? "Criar" : "Editar"}
+        </Button>
       </div>
 
-      {current && (
-        <div className="space-y-4">
-          <div className="space-y-2">
-            <Label htmlFor="cc-name">Nome</Label>
-            <Input
-              id="cc-name"
-              value={current.name}
-              onChange={(e) =>
-                setCurrent({ ...current, name: e.target.value })
-              }
-              placeholder="Nome do componente"
-            />
-          </div>
-          <div className="space-y-2">
-            <Label htmlFor="cc-html">HTML</Label>
-            <Textarea
-              id="cc-html"
-              rows={3}
-              value={current.html}
-              onChange={(e) =>
-                setCurrent({ ...current, html: e.target.value })
-              }
-            />
-          </div>
-          <div className="space-y-2">
-            <Label htmlFor="cc-css">CSS</Label>
-            <Textarea
-              id="cc-css"
-              rows={3}
-              value={current.css}
-              onChange={(e) => setCurrent({ ...current, css: e.target.value })}
-            />
-          </div>
-          <div className="space-y-2">
-            <Label htmlFor="cc-js">JS</Label>
-            <Textarea
-              id="cc-js"
-              rows={3}
-              value={current.js}
-              onChange={(e) => setCurrent({ ...current, js: e.target.value })}
-            />
-          </div>
-          <Button size="sm" onClick={handleSave}>
-            {isNew ? "Criar" : "Atualizar"} componente
-          </Button>
-        </div>
-      )}
+      <Dialog open={open} onOpenChange={setOpen}>
+        <DialogContent className="sm:max-w-lg">
+          <DialogHeader>
+            <DialogTitle>{isNew ? "Novo Componente" : "Editar Componente"}</DialogTitle>
+          </DialogHeader>
+          {current && (
+            <div className="space-y-4">
+              <div className="space-y-2">
+                <Label htmlFor="cc-name">Nome</Label>
+                <Input
+                  id="cc-name"
+                  value={current.name}
+                  onChange={(e) =>
+                    setCurrent({ ...current, name: e.target.value })
+                  }
+                  placeholder="Nome do componente"
+                />
+              </div>
+              <div className="space-y-2">
+                <Label htmlFor="cc-html">HTML</Label>
+                <Textarea
+                  id="cc-html"
+                  rows={3}
+                  value={current.html}
+                  onChange={(e) =>
+                    setCurrent({ ...current, html: e.target.value })
+                  }
+                />
+              </div>
+              <div className="space-y-2">
+                <Label htmlFor="cc-css">CSS</Label>
+                <Textarea
+                  id="cc-css"
+                  rows={3}
+                  value={current.css}
+                  onChange={(e) =>
+                    setCurrent({ ...current, css: e.target.value })
+                  }
+                />
+              </div>
+              <div className="space-y-2">
+                <Label htmlFor="cc-js">JS</Label>
+                <Textarea
+                  id="cc-js"
+                  rows={3}
+                  value={current.js}
+                  onChange={(e) =>
+                    setCurrent({ ...current, js: e.target.value })
+                  }
+                />
+              </div>
+              <Button size="sm" onClick={handleSave}>
+                {isNew ? "Criar" : "Atualizar"} componente
+              </Button>
+            </div>
+          )}
+        </DialogContent>
+      </Dialog>
     </div>
   );
 }
